\name{GEWIST-package}
\alias{GEWIST-package}
\alias{GEWIST}
\docType{package}
\title{
Gene Environment Wide Interaction Search Threshold }
\description{
This 'GEWIST' package provides statistical tools to efficiently optimize SNP prioritization for gene-gene and gene-environment interactions.
}
\details{
\tabular{ll}{
Package: \tab PathWei\cr
Type: \tab Package\cr
Version: \tab 0.99.z\cr
License: \tab GPL-2 \cr
LazyLoad: \tab yes\cr
}
}
\author{
Wei Q. Deng <dengwq@mcmaster.ca>
Guillaume Pare <pareg@mcmaster.ca>
}
\references{
 Deng W.Q, Pare G. (2011) 
 A fast algorithm to optimize SNP prioritization for gene-gene and gene-environment interactions. 
 \emph{Genetic Epidemiology}. 35: 729-738. doi: 10.1002/gepi.20624

 Pare G, Cook NR, Ridker PM, Chasman DI (2010) 
 On the Use of Variance per Genotype as a Tool to Identify Quantitative Trait Interaction Effects: A Report 
 from the Women's Genome Health Study. 
 \emph{PLoS Genet} 6(6): e1000981. doi:10.1371/journal.pgen.1000981

 Levene H. (1960) 
 Robust tests for equality of variances. In \emph{Contributions to Probability and Statistics: Essays in Honor of Harold 
 Hotelling} 
 eds:I. Olkin, S.G. Ghurye, W. Hoeffding, W.G. Madow & H.B.Mann, pp.278-292. Stanford: Stanford University Press.
}

\keyword{ genetic interactions }
\keyword{ quantitative trait }
\keyword{ variance prioritization }% __ONLY ONE__ keyword per line
\keyword{ Levene's Test}
\keyword{ Bonferroni correction}
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
